Move mockAuth.server to /mocks directory
